+++ unique_toolkit/CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.12] - 2024-08-7
+- added `completedAt` datetime to `unique_sdk.Message.modify` and `unique_sdk.Message.modify_async`
+
 ## [0.5.11] - 2024-08-6
 - made all domain specific functions and classes directly importable from `unique_toolkit.[DOMAIN_NAME]`
 - renamed `RerankerConfig` to `ContentRerankerConfig`